+++ src/components/authentication/signup/footer/AuthFooter.tsx
